+++ data-manager/api/src/main/java/org/orbisgis/orbisdata/datamanager/api/dataset/ISpatialTable.java
@@ -36,7 +36,6 @@
  */
 package org.orbisgis.orbisdata.datamanager.api.dataset;
 
-import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.Geometry;
 import org.orbisgis.commons.annotations.NotNull;
 import org.orbisgis.commons.annotations.Nullable;
@@ -205,15 +204,4 @@ public interface ISpatialTable<T, U> extends ITable<T, U> {
      */
     @Nullable
     ISpatialTable<T, U> reproject(int srid);
-
-    @Override
-    ISpatialTable<T, U> filter(String filter);
-
-    @NotNull
-    @Override
-    ISpatialTable<T, U> columns(@NotNull String... columns);
-
-    @NotNull
-    @Override
-    ISpatialTable<T, U> columns(@NotNull List<String> columns);
 }