+++ data-manager/jdbc/src/main/java/org/orbisgis/orbisdata/datamanager/jdbc/JdbcSpatialTable.java
@@ -40,7 +40,6 @@ import org.h2gis.utilities.GeometryTableUtilities;
 import org.h2gis.utilities.SpatialResultSet;
 import org.h2gis.utilities.SpatialResultSetMetaData;
 import org.h2gis.utilities.Tuple;
-import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.Geometry;
 import org.orbisgis.commons.annotations.NotNull;
 import org.orbisgis.commons.annotations.Nullable;
@@ -48,7 +47,6 @@ import org.orbisgis.orbisdata.datamanager.api.dataset.DataBaseType;
 import org.orbisgis.orbisdata.datamanager.api.dataset.IJdbcSpatialTable;
 import org.orbisgis.orbisdata.datamanager.api.dataset.IRaster;
 import org.orbisgis.orbisdata.datamanager.api.datasource.IJdbcDataSource;
-import org.orbisgis.orbisdata.datamanager.jdbc.dsl.WhereBuilder;
 import org.orbisgis.orbisdata.datamanager.jdbc.resultset.StreamResultSet;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -399,34 +397,6 @@ public abstract class JdbcSpatialTable extends JdbcTable implements IJdbcSpatial
         } catch (SQLException e) {
             LOGGER.error("Unable to get the metadata.", e);
             return null;
-        }
-    }
-
-    @Override
-    @Nullable
-    public JdbcSpatialTable filter(String filter) {
-        return (JdbcSpatialTable)where(filter).getSpatialTable();
-    }
-
-    @Override
-    @NotNull
-    public JdbcSpatialTable columns(@NotNull String... columns) {
-        WhereBuilder builder = new WhereBuilder(getQuery(columns), getJdbcDataSource());
-        if (isSpatial()) {
-            return (JdbcSpatialTable) builder.getSpatialTable();
-        } else {
-            throw new IllegalArgumentException("A ISpatialTable should keep at least on spatial field.");
-        }
-    }
-
-    @Override
-    @NotNull
-    public JdbcSpatialTable columns(@NotNull List<String> columns) {
-        WhereBuilder builder = new WhereBuilder(getQuery(columns.toArray(new String[0])), getJdbcDataSource());
-        if (isSpatial()) {
-            return (JdbcSpatialTable) builder.getSpatialTable();
-        } else {
-            throw new IllegalArgumentException("A ISpatialTable should keep at least on spatial field.");
         }
     }
 }