+++ data-manager/jdbc/src/main/java/org/orbisgis/orbisdata/datamanager/jdbc/dsl/BuilderResult.java
@@ -37,26 +37,14 @@
 package org.orbisgis.orbisdata.datamanager.jdbc.dsl;
 
 import groovy.lang.Closure;
-import org.h2gis.utilities.wrapper.ConnectionWrapper;
-import org.h2gis.utilities.wrapper.StatementWrapper;
 import org.orbisgis.commons.annotations.NotNull;
 import org.orbisgis.commons.printer.ICustomPrinter;
 import org.orbisgis.orbisdata.datamanager.api.dataset.ISpatialTable;
 import org.orbisgis.orbisdata.datamanager.api.dataset.ITable;
 import org.orbisgis.orbisdata.datamanager.api.datasource.IJdbcDataSource;
 import org.orbisgis.orbisdata.datamanager.api.dsl.IBuilderResult;
-import org.orbisgis.orbisdata.datamanager.jdbc.JdbcDataSource;
-import org.orbisgis.orbisdata.datamanager.jdbc.h2gis.H2gisSpatialTable;
-import org.orbisgis.orbisdata.datamanager.jdbc.h2gis.H2gisTable;
-import org.orbisgis.orbisdata.datamanager.jdbc.postgis.PostgisSpatialTable;
-import org.orbisgis.orbisdata.datamanager.jdbc.postgis.PostgisTable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
 
 /**
  * Implementation of {@link IBuilderResult}.
@@ -98,43 +86,10 @@ public abstract class BuilderResult implements IBuilderResult {
                 return table.asType(clazz);
             }
         }
-        Statement statement;
-        Connection con;
-        try {
-            con = getDataSource().getConnection();
-            if(con == null){
-                LOGGER.error("Unable to create the StatementWrapper.\n");
-                return null;
-            }
-            statement = con.createStatement(ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_UPDATABLE);
-        } catch (SQLException e) {
-            LOGGER.error("Unable to create the StatementWrapper.\n", e);
-            return null;
-        }
-        switch (getDataSource().getDataBaseType()) {
-            default:
-            case H2GIS:
-                if (!(statement instanceof StatementWrapper)) {
-                    LOGGER.warn("The statement class not compatible with the database.");
-                    statement = new StatementWrapper(statement, new ConnectionWrapper(con));
-                }
-                if (ISpatialTable.class.isAssignableFrom(clazz)) {
-                    return new H2gisSpatialTable(null, getQuery(), statement, getDataSource());
-                } else if (ITable.class.isAssignableFrom(clazz)) {
-                    return new H2gisTable(null, getQuery(), statement, getDataSource());
-                }
-            case POSTGIS:
-                if (!(statement instanceof org.h2gis.postgis_jts.StatementWrapper)) {
-                    LOGGER.error("The statement class not compatible with the database.");
-                    break;
-                }
-                if (ISpatialTable.class.isAssignableFrom(clazz)) {
-                    return new PostgisSpatialTable(null, getQuery(),
-                            (org.h2gis.postgis_jts.StatementWrapper) statement, getDataSource());
-                } else if (ITable.class.isAssignableFrom(clazz)) {
-                    return new PostgisTable(null, getQuery(),
-                            (org.h2gis.postgis_jts.StatementWrapper) statement, getDataSource());
-                }
+        if (ISpatialTable.class.isAssignableFrom(clazz)) {
+            return getSpatialTable();
+        } else if (ITable.class.isAssignableFrom(clazz)) {
+            return getTable();
         }
         return null;
     }
@@ -142,16 +97,16 @@ public abstract class BuilderResult implements IBuilderResult {
     @NotNull
     @Override
     public String toString() {
-        return getQuery();
+        return "(" + getQuery() + ")";
     }
 
     @Override
     public ITable getTable() {
-        return (ITable) this.asType(ITable.class);
+        return getDataSource().getTable(toString());
     }
 
     @Override
     public ISpatialTable getSpatialTable() {
-        return (ISpatialTable) this.asType(ISpatialTable.class);
+        return getDataSource().getSpatialTable(toString());
     }
 }