+++ process-manager/process/src/main/java/org/orbisgis/orbisdata/processmanager/process/GroovyProcessManager.java
@@ -44,8 +44,6 @@ import org.orbisgis.orbisdata.processmanager.api.IProcess;
 import org.orbisgis.orbisdata.processmanager.api.IProcessBuilder;
 import org.orbisgis.orbisdata.processmanager.api.IProcessFactory;
 import org.orbisgis.orbisdata.processmanager.api.IProcessManager;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.util.List;
 import java.util.Map;